Refactoring and dedup
